Reorganized `...Module.kt` and Serializer files. `ComposeDestinations` now don't require the generated code's path to be included manually (since Kotlin 1.8.0), so code for that got removed.
