--- app/src/main/java/illyan/jay/data/network/model/GeoPointParceler.kt
@@ -1,32 +1,0 @@
-/*
- * Copyright (c) 2023 Balázs Püspök-Kiss (Illyan)
- *
- * Jay is a driver behaviour analytics app.
- *
- * This file is part of Jay.
- *
- * Jay is free software: you can redistribute it and/or modify it under the
- * terms of the GNU General Public License as published by the Free Software
- * Foundation, either version 3 of the License, or (at your option) any later version.
- * Jay is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
- * without even the implied warranty of MERCHANTABILITY or FITNESS FOR
- * A PARTICULAR PURPOSE. See the GNU General Public License for more details.
- *
- * You should have received a copy of the GNU General Public License along with Jay.
- * If not, see <https://www.gnu.org/licenses/>.
- */
-
-package illyan.jay.data.network.model
-
-import android.os.Parcel
-import com.google.firebase.firestore.GeoPoint
-import kotlinx.parcelize.Parceler
-
-object GeoPointParceler : Parceler<GeoPoint> {
-    override fun create(parcel: Parcel) = GeoPoint(parcel.readDouble(), parcel.readDouble())
-
-    override fun GeoPoint.write(parcel: Parcel, flags: Int) {
-        parcel.writeDouble(latitude)
-        parcel.writeDouble(longitude)
-    }
-}
