+++ app/src/main/java/illyan/jay/di/RoomModule.kt
@@ -16,7 +16,7 @@
  * If not, see <https://www.gnu.org/licenses/>.
  */
 
-package illyan.jay.data.disk
+package illyan.jay.di
 
 import android.content.Context
 import androidx.room.Room
@@ -27,6 +27,7 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
+import illyan.jay.data.disk.JayDatabase
 import timber.log.Timber
 import javax.inject.Singleton
