+++ app/src/main/java/illyan/jay/util/TestAdUnitIds.kt
@@ -1,0 +1,33 @@
+/*
+ * Copyright (c) 2023 Balázs Püspök-Kiss (Illyan)
+ *
+ * Jay is a driver behaviour analytics app.
+ *
+ * This file is part of Jay.
+ *
+ * Jay is free software: you can redistribute it and/or modify it under the
+ * terms of the GNU General Public License as published by the Free Software
+ * Foundation, either version 3 of the License, or (at your option) any later version.
+ * Jay is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY;
+ * without even the implied warranty of MERCHANTABILITY or FITNESS FOR
+ * A PARTICULAR PURPOSE. See the GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License along with Jay.
+ * If not, see <https://www.gnu.org/licenses/>.
+ */
+
+package illyan.jay.util
+
+/**
+ * Test Ad Unit IDs from [https://developers.google.com/admob/android/test-ads](https://developers.google.com/admob/android/test-ads)
+ */
+object TestAdUnitIds {
+    const val AppOpen = "ca-app-pub-3940256099942544/3419835294"
+    const val Banner = "ca-app-pub-3940256099942544/6300978111"
+    const val Interstitial = "ca-app-pub-3940256099942544/1033173712"
+    const val InterstitialVideo = "ca-app-pub-3940256099942544/8691691433"
+    const val Rewarded = "ca-app-pub-3940256099942544/5224354917"
+    const val RewardedInterstitial = "ca-app-pub-3940256099942544/5354046379"
+    const val NativeAdvanced = "ca-app-pub-3940256099942544/2247696110"
+    const val NativeAdvancedVideo = "ca-app-pub-3940256099942544/1044960115"
+}
